JASPER-43 #comment JMP Discovery Client for Jasper
Commit to include the JMP Client and the setup for it.

+++ jasper-platform/zipRoot/setup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+CDIR=`pwd`
+
 function setup_jsb {
 mkdir -p logs
 cd jsb-core
@@ -9,12 +11,29 @@ chmod 755 jsbAutoStart
 chmod 600 config/jsb.jmxremote.password
 mv wrapper.jar libs/
 mkdir activemq-data
-cd exec
-ln -s ../activemq-data .
-cd ..
+if [ ! -L exec/activemq-data ]
+  then  
+    cd exec
+    ln -s ../activemq-data .
+    cd ..
+fi    
+}
+
+function setup_jmc_discovery {
+  if [ -e jsb-core/jmp.tar.gz ]; then
+    cd $CDIR
+    cd ..
+    cp $CDIR/jsb-core/jmp.tar.gz .
+    tar -xzf jmp.tar.gz
+    rm jmp.tar.gz
+    cd $CDIR
+  fi  
 }
 
 function setup_jta {
+if [ ! -e ../jmp ]; then
+  setup_jmc_discovery
+fi  
 mkdir -p JTAs
 mkdir -p logs
 cd jsb-core
@@ -26,10 +45,12 @@ if ! [ -d mule-standalone-3.3.0 ]; then
    mv default mule-standalone-3.3.0/apps/
    rm default.tar.gz
 fi
-
 }
 
 function setup_all {
+if [ ! -e ../jmp ]; then
+  setup_jmc_discovery
+fi
 mkdir -p JTAs
 mkdir -p logs
 cd jsb-core
@@ -47,9 +68,12 @@ chmod 755 exec/wrapper*
 chmod 755 jsbAutoStart
 chmod 755 jtaAutoStart
 mkdir activemq-data
-cd exec
-ln -s ../activemq-data .
-cd ..
+if [ ! -L exec/activemq-data ]
+  then  
+    cd exec
+    ln -s ../activemq-data .
+    cd ..
+fi
 }
 
 case "$1" in
